+++ operator_ui/support/factories/jobSpecV2.ts
@@ -5,6 +5,7 @@ import {
   FluxMonitorJobV2Spec,
   KeeperV2Spec,
   OffChainReportingOracleJobV2Spec,
+  CronV2Spec,
 } from 'core/store/models'
 import { generateUuid } from '../test-helpers/generateUuid'
 
@@ -46,6 +47,7 @@ export function ocrJobSpecV2(
     fluxMonitorSpec: null,
     directRequestSpec: null,
     keeperSpec: null,
+    cronSpec: null,
     errors: [],
     maxTaskDuration: '',
     pipelineSpec: {
@@ -81,6 +83,7 @@ export function fluxMonitorJobV2(
     directRequestSpec: null,
     offChainReportingOracleSpec: null,
     keeperSpec: null,
+    cronSpec: null,
     fluxMonitorSpec,
     errors: [],
     maxTaskDuration: '',
@@ -118,6 +121,7 @@ export function directRequestJobV2(
     offChainReportingOracleSpec: null,
     keeperSpec: null,
     fluxMonitorSpec: null,
+    cronSpec: null,
     errors: [],
     maxTaskDuration: '',
     pipelineSpec: {
@@ -147,6 +151,35 @@ export function keeperJobV2(
     keeperSpec,
     offChainReportingOracleSpec: null,
     fluxMonitorSpec: null,
+    cronSpec: null,
+    errors: [],
+    maxTaskDuration: '',
+    pipelineSpec: {
+      dotDagSource: '',
+    },
+  }
+}
+
+export function cronJobV2(
+  spec: Partial<CronV2Spec['cronSpec']> = {},
+  config: Partial<{
+    name?: string
+    id?: string
+  }> = {},
+): JobSpecV2 {
+  const cronSpec = partialAsFull<CronV2Spec['cronSpec']>({
+    createdAt: spec.createdAt || new Date(1600775300410).toISOString(),
+  })
+  return {
+    name: config.name || 'Cron V2 job',
+    type: 'cron',
+    schemaVersion: 1,
+    directRequestSpec: null,
+    keeperSpec: null,
+    offChainReportingOracleSpec: null,
+    fluxMonitorSpec: null,
+    cronSpec,
+    webSpec: null,
     errors: [],
     maxTaskDuration: '',
     pipelineSpec: {